requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.0
binaryornot==0.4.4
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
cron-descriptor==1.4.5
Django==5.1
django-browser-reload==1.15.0
django-htmx==1.19.0
django-tailwind==3.8.0
django-timezone-field==7.0
django4-background-tasks==1.2.10
Faker==28.0.0
fontawesome-free==5.15.4
fontawesomefree==6.6.0
idna==3.8
Jinja2==3.1.4
kombu==5.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
prompt_toolkit==3.0.47
Pygments==2.18.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.8.0
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
